Seperate the API routes into own file
